splitted providers into core and extension providers
