+++ src/generators/crud/providers/extensions/DateTimeProvider.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace schmunk42\giiant\generators\crud\providers;
+namespace schmunk42\giiant\generators\crud\providers\extensions;
 
 class DateTimeProvider extends \schmunk42\giiant\base\Provider
 {
